+++ FFT/AudioProcesisng/wavread.h
@@ -123,7 +123,7 @@ class WavRead
 
 		
 		/// A buffer tartalmat a kivalasztott csatornabol felirja komplex alakban [-1..1] intervallumon
-		void fillBufferComplex(double* buffer, channel_t channel, int len = 0, int is32k=0);
+		void fillBufferComplex(float* buffer, channel_t channel, int len = 0, int is32k=0);
 
 		/// Buffer tartalmat kiirja raw formaban
 		void fillBuffer(int size, int offset, void* buffer);
@@ -188,11 +188,11 @@ class WavWrite{
 		~WavWrite();
 
 		/// hibas
-		void writeComplex_old(double *data, int length);
+		void writeComplex_old(float *data, int length);
 		
 		/// kirja a buffert - hibas
 		//void write(int *left, int *right, int length);
-		void write(double *left, double *right, void* workbuffer, int length, int is32k = 0);
+		void write(float *left, float *right, void* workbuffer, int length, int is32k = 0);
 };
 
 #endif // WAVREAD_H
